Big warning cleanup everywhere, plus some changes to yacc options.

+++ caj_vm_asm.h
@@ -172,7 +172,7 @@ public:
     assert(funcs[cfunc->func_num] == cfunc);
 
     printf("DEBUG: assembler: starting function %s (number %i) @ %i\n", 
-	   cfunc->name, cfunc->func_num, bytecode.size());
+	   cfunc->name, cfunc->func_num, (int)bytecode.size());
 
     vm_function *func = funcs[cfunc->func_num];
     func->insn_ptr = func_start = bytecode.size();
